settings component refactor
